Shuffling logic to transformProps, making Typescript happy.
