+++ plugins/plugin-chart-echarts/src/types.ts
@@ -1,4 +1,5 @@
 import { DataRecord } from '@superset-ui/chart';
+import ECharts from 'echarts';
 
 export type TimestampType = string | number | Date;
 
@@ -9,3 +10,21 @@ export interface EchartsTimeseriesRawDatum extends DataRecord {
 export interface EchartsTimeseriesDatum extends DataRecord {
   __timestamp: Date;
 }
+
+export type EchartsLineProps = {
+  area: number;
+  colorScheme: string;
+  contributionMode?: string;
+  height: number;
+  seriesType: string;
+  logAxis: boolean;
+  width: number;
+  stack: boolean;
+  markerEnabled: boolean;
+  markerSize: number;
+  minorSplitLine: boolean;
+  opacity: number;
+  data: EchartsTimeseriesDatum[]; // please add additional typing for your data here
+  // add typing here for the props you pass in from transformProps.ts!
+  echartOptions: ECharts.EChartOption;
+};
